refactor: extract FavoriteProjectScreen composables into components package
