+++ app/src/main/java/com/example/twdist_android/features/favorite/presentation/components/FavoriteProjectsLifecycleEffect.kt
@@ -1,0 +1,25 @@
+package com.example.twdist_android.features.favorite.presentation.components
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
+import androidx.lifecycle.LifecycleOwner
+
+@Composable
+internal fun FavoriteProjectsLifecycleEffect(
+    lifecycleOwner: LifecycleOwner,
+    onLoadProjects: () -> Unit
+) {
+    DisposableEffect(lifecycleOwner) {
+        val observer = LifecycleEventObserver { _, event ->
+            if (event == Lifecycle.Event.ON_RESUME) {
+                onLoadProjects()
+            }
+        }
+        lifecycleOwner.lifecycle.addObserver(observer)
+        onDispose {
+            lifecycleOwner.lifecycle.removeObserver(observer)
+        }
+    }
+}
