+++ app/src/main/java/com/example/twdist_android/features/favorite/presentation/components/InlineErrorMessage.kt
@@ -1,0 +1,17 @@
+package com.example.twdist_android.features.favorite.presentation.components
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+internal fun InlineErrorMessage(message: String) {
+    Text(
+        text = message,
+        color = MaterialTheme.colorScheme.error,
+        modifier = Modifier.padding(16.dp)
+    )
+}
